Add external modules required for aarch64 environment to mozc-deps.yaml
To obtain information about external modules required for the aarch64 environment, particularly regarding cpython,
we added bazel execution using the qemu-user-static and binfmt environment and aarch64 emulation through flatpak.
(with qemu-user-static and support-binfmt packages on Ubuntu)
(with qemu-user-static-binfmt package on Arch)

For execution, the qemu-user-static-binfmt environment and installation of the following flatpak packages are necessary.
  org.fcitx.Fcitx5/aarch64/stable
  org.kde.Sdk/aarch64/6.6
  org.freedesktop.Sdk/aarch64/23.08
  org.freedesktop.Sdk.Extension.bazel/aarch64/23.08
  org.freedesktop.Sdk.Extension.llvm18/aarch64/23.08

Additionally, I implemented a process to add the 'only-arches' item if the URL contains 'cpython' and ['x86_64' or 'aarch64'].

Note:
  org.kde.Sdk//6.6 or org.kde.Sdk//6.7 and org.freedesktop.Sdk//23.08 are interdependent.
  As the bazel version in org.freedesktop.Sdk.Extension.bazel//23.08 has been updated to 7.3.2, it can be used for building with bzlmod enabled.
  In the future, it seems that when org.kde.Sdk//6.8 becomes available, org.freedesktop.Sdk//24.08 will be used.